requirements.txt
South==0.7.4
sorl-thumbnail==11.12
locast-core==1.0
django-social-auth==0.7.7
